# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is api for beacon services",
        "title": "Swagger API",
        "contact": {
            "name": "API Support",
            "email": "paradoxfm@mail.ru"
        },
        "license": {},
        "version": "1.0"
    },
    "paths": {
...
        "/users/me": {
            "get": {
                "description": "Getting a user profile",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Profile request",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.ProfileResponse"
                        }
                    },
                    "400": {
                        "description": "err"
                    },
                    "500": {
                        "description": "err"
                    }
                }
            }
        },
        "/users/me/avatar": {
            "put": {
                "description": "Update user avatar (jpeg 250x250)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update user avatar",
                "parameters": [
                    {
                        "description": "avatar",
                        "name": "avatar",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.AvatarIdResponse"
                        }
                    },
                    "400": {
                        "description": "err"
                    },
                    "500": {
                        "description": "err"
                    }
                }
            }
        },
        "/users/me/password": {
            "put": {
                "description": "Change user password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Change password",
                "parameters": [
                    {
                        "description": "Password change request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "400": {
                        "description": "err"
                    },
                    "500": {
                        "description": "err"
                    }
                }
            }
        },
        "/users/me/push": {
            "put": {
                "description": "Update user push id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update user push id",
                "parameters": [
                    {
                        "description": "Request with push id",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.UpdatePushRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "400": {
                        "description": "err"
                    },
                    "500": {
                        "description": "err"
                    }
                }
            }
        },
        "/users/password/reset": {
            "post": {
                "description": "Reset user password by email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Password reset",
                "parameters": [
                    {
                        "description": "Password reset request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.ResetPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "400": {
                        "description": "err"
                    },
                    "500": {
                        "description": "err"
                    }
                }
            }
        },
        "/users/signUp": {
            "post": {
                "description": "User registration email + password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "User registration",
                "parameters": [
                    {
                        "description": "Registration Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.RegistrationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "400": {
                        "description": "err"
                    },
                    "500": {
                        "description": "err"
                    }
                }
            }
        },
        "/zone/all": {
            "get": {
                "description": "Getting a list of zones for an authorized user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Zones"
                ],
                "summary": "Zone list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.GeoZoneResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "err"
                    },
                    "500": {
                        "description": "err"
                    }
                }
            }
        },
        "/zone/delete/{id}": {
            "delete": {
                "description": "Deleting a zone for an authorized user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Zones"
                ],
                "summary": "Deleting a zone",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id zone",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "400": {
                        "description": "err"
                    },
                    "500": {
                        "description": "err"
                    }
                }
            }
        },
        "/zone/find/{id}": {
            "get": {
                "description": "Loading zone for authorized user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Zones"
                ],
                "summary": "Loading zone",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id zone",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.GeoZoneResponse"
                        }
                    },
                    "400": {
                        "description": "err"
                    },
                    "500": {
                        "description": "err"
                    }
                }
            }
        },
        "/zone/save": {
            "post": {
                "description": "Creating a zone for an authorized user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Zones"
                ],
                "summary": "Creating a zone",
                "parameters": [
                    {
                        "description": "Request with zone settings",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.ZoneCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.GeoZoneResponse"
                        }
                    },
                    "400": {
                        "description": "err"
                    },
                    "500": {
                        "description": "err"
                    }
                }
            }
        },
        "/zone/snap/{id}": {
            "put": {
                "description": "Linking the zone to the tracker list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Zones"
                ],
                "summary": "Zone binding",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id зоны",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request linking trackers to a zone",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.ZoneSnapRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "400": {
                        "description": "err"
                    },
                    "500": {
                        "description": "err"
                    }
                }
            }
        },
        "/zone/update/{id}": {
            "put": {
                "description": "Updating zone settings for an authorized user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Zones"
                ],
                "summary": "Zone update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id зоны",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request with zone settings",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.ZoneCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "400": {
                        "description": "err"
                    },
                    "500": {
                        "description": "err"
                    }
                }
            }
        }
    },
    "definitions": {
        "model.AvatarIdResponse": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
...
        "model.ChangePasswordRequest": {
            "type": "object",
            "properties": {
                "new_password": {
                    "type": "string"
                },
                "old_password": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "model.GeoZoneResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "points": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ZonePoint"
                    }
                },
                "trackers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "model.ProfileResponse": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "trackers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UserTracker"
                    }
                }
            }
        },
        "model.RegistrationRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "сonfirm": {
                    "type": "string"
                }
            }
        },
        "model.ResetPasswordRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdatePushRequest": {
            "type": "object",
            "properties": {
                "push_id": {
                    "type": "string"
                }
            }
        },
        "model.UserTracker": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "offset": {
                    "type": "integer"
                }
            }
        },
        "model.ZoneCreateRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "points": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ZonePoint"
                    }
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "model.ZonePoint": {
            "type": "object",
            "properties": {
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                }
            }
        },
        "model.ZoneSnapRequest": {
            "type": "object",
            "properties": {
                "trackers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}